requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.2
defusedxml==0.7.1
Django==5.2
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
idna==3.10
iniconfig==2.1.0
kombu==5.5.2
oauthlib==3.2.2
packaging==25.0
pluggy==1.5.0
prompt_toolkit==3.0.50
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
